+++ probe-rs/src/lib.rs
@@ -91,7 +91,7 @@ mod test;
 
 pub use crate::config::{CoreType, InstructionSet, Target};
 pub use crate::core::{
-    core_dump::{CoreDump, CoreDumpError},
+    dump::{CoreDump, CoreDumpError},
     exception_handler_for_core, Architecture, BreakpointCause, Core, CoreInformation,
     CoreInterface, CoreRegister, CoreRegisters, CoreState, CoreStatus, HaltReason,
     MemoryMappedRegister, RegisterId, RegisterRole, RegisterValue, SemihostingCommand,
